EntityPool updated to use object pooling for entities and components.
This is a breaking change. There are new methods for setting, getting and
removing components on entities, including one method (addComponent) for
generating a new component object from a component pool and one (setComponent)
for setting a primitive value as a component on an entity.

The benefit of this change is reduced garbage and garbage collection.